fix import name in example and rename file
