+++ src/types/CogsPluginManifest.ts
